Report CPU Utilization for the rNode process
